[#22] rename: components 폴더명 수정 및 파일 위치 변경 kakaomap -> stampMap
