+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import localFont from 'next/font/local';
 import './globals.css';
 import RQProviders from '@/providers/RQRovider';
-import KakaoMapLoader from '@/components/kakaomap/KakaoMapLoader';
+import KakaoMapLoader from '@/components/stampMap/KakaoMapLoader';
 
 const geistSans = localFont({
   src: './fonts/GeistVF.woff',
@@ -17,8 +17,8 @@ const geistMono = localFont({
 });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app'
+  title: '모아 메인 페이지',
+  description: '모아 메인 페이지 입니다.'
 };
 
 export default function RootLayout({
