+++ src/components/stampMap/KakaoMapLoader.tsx
