+++ src/components/stampMap/ScrollButtonSwiper.tsx
